Address review comments.

+++ sdk-base/src/main/java/com/mapbox/maps/extension/observable/model/TileID.kt
@@ -10,15 +10,15 @@ data class TileID(
    * The zoom level.
    */
   @SerializedName("z")
-  val zoom: Double,
+  val zoom: Long,
   /**
    * The x coordinate of the tile
    */
   @SerializedName("x")
-  val x: Double,
+  val x: Long,
   /**
    * The y coordinate of the tile
    */
   @SerializedName("y")
-  val y: Double
+  val y: Long
 )
